chore: re-organize project directories
