+++ src/writer/metadata.go
